rename spark sent modal
