+++ src/app/app.module.ts
@@ -36,7 +36,7 @@ import { CreateAccountComponent } from './screens/create-account/create-account.
 import { ForgotPasswordComponent } from './screens/forgot-password/forgot-password.component';
 import { MainMenuComponent } from './_components/main-menu/main-menu.component';
 import { NotificationComponent } from './_components/notification/notification.component';
-import { ShareSparkModalComponent } from './_components/share-spark-modal/share-spark-modal.component';
+import { ShareSparkModalComponent } from './_components/spark-sent-modal/spark-sent-modal.component';
 import { TeamDashboardComponent } from './screens/team-dashboard/team-dashboard.component';
 import { ChartModule } from 'angular-highcharts';
 import { SendThanksModalComponent } from './_components/send-thanks-modal/send-thanks-modal.component'; // https://api.highcharts.com/highcharts/
